pyproject.toml
[project]
authors = [
	{name = "Daniil Shomin", email = "d.sh.v@yandex.ru"},
]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.7",
    "validators>=0.34.0",
]
